+++ src/app/Playground/AppShell.xaml.cs
@@ -1,6 +1,6 @@
 ﻿namespace Playground;
 
-public partial class AppShell : Shell
+public partial class AppShell
 {
     public AppShell()
     {